+++ commands/owner/yearPoll.ts
@@ -1,7 +1,7 @@
 import chalk from "chalk";
 import { MessageEmbed, MessageActionRow, MessageSelectMenu } from "discord.js";
 import { ICommand } from "wokcommands";
-import { checkForRoles } from "../../utils/roleUtils";
+import { checkForRoles } from "../../utils/roles";
 
 export default {
   name: "yearPoll",
