+++ index.ts
@@ -4,10 +4,10 @@ import WOKCommands from "wokcommands";
 import path from "path";
 import chalk from "chalk";
 import dotenv from "dotenv";
-import { isDocker } from "./utils/util";
+import { isDocker } from "./utils/docker";
 
 // import custom modules
-import { checkForRoles, checkIfCollectionsExist } from "./utils/roleUtils";
+import { checkForRoles, checkIfCollectionsExist } from "./utils/roles";
 import { classModel } from "./models/classModel";
 import { staffModel } from "./models/staffModel";
 import { yearModel } from "./models/yearModel";
